CLOUD-Project/vehicle_data_producer/requirements.txt
PySide6
paho-mqtt<2.0.0
